+++ kite/src/main/kotlin/com/tang/kite/session/SqlSession.kt
@@ -20,9 +20,9 @@ interface SqlSession : AutoCloseable {
 
     fun <T> insertSelective(method: Method, mapperInterface: Class<T>, parameter: Any): Int
 
-    fun <T> batchInsert(method: Method, mapperInterface: Class<T>, parameter: Any): Int
+    fun <T> batchInsert(method: Method, mapperInterface: Class<T>, parameter: Any, batchSize: Int): Int
 
-    fun <T> batchInsertSelective(method: Method, mapperInterface: Class<T>, parameter: Any): Int
+    fun <T> batchInsertSelective(method: Method, mapperInterface: Class<T>, parameter: Any, batchSize: Int): Int
 
     fun <T> update(method: Method, mapperInterface: Class<T>, parameter: Any): Int